move educator module to classes in handlers
