+++ spec/handlers/complete_educator_profile_spec.rb
