+++ spec/handlers/verify_email_by_pin_spec.rb
